Pulling sensor values from OHM.

+++ LCDHardwareMonitor/src/renderer_d3d11.hpp
@@ -734,7 +734,7 @@ TeardownRenderer(D3DRendererState* s)
 DrawCall*
 PushDrawCall(D3DRendererState* s)
 {
-	Assert(s->drawCallCount < ArrayCount(s->drawCalls));
+	Assert(s->drawCallCount < (u32) ArrayCount(s->drawCalls));
 
 	DrawCall* drawCall = &s->drawCalls[s->drawCallCount++];
 	return drawCall;
